[Refactoring ⚙️] 리뷰 받은 내용 수정
