+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,6 +1,6 @@
 import { Box, Center, Divider, Flex } from "@chakra-ui/react"
 
-import LogoButton from "@components/LogoButton/LogoButton"
+import LogoButton from "@components/LogoLink/LogoLink"
 
 import BackButton from "./components/BackButton"
 import CreateAccount from "./components/CreateAccount"
